Enhance AI receptionist functionality and authentication flow
- Introduced `ensureFreshAccessToken` to streamline token validation and refresh logic.
- Updated the AI receptionist to handle session end events and manage call termination based on user intent.
- Improved error handling and user feedback for authentication issues.
- Added utility functions for message normalization and date resolution in booking context.
- Enhanced the onboarding process with better session management and token handling.

+++ apps/server/src/modules/integrations/integration.service.ts
@@ -153,8 +153,8 @@ function getDayKey(date: Date, timezone: string): string {
 function formatSlotLabel(startIso: string, timezone: string): string {
   return new Intl.DateTimeFormat('en-US', {
     timeZone: timezone,
-    weekday: 'short',
-    month: 'short',
+    weekday: 'long',
+    month: 'long',
     day: 'numeric',
     hour: 'numeric',
     minute: '2-digit',
